requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
feedparser==6.0.11
groq==0.15.0
pydantic==2.10.5
apscheduler==3.10.4
aiosqlite==0.20.0
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
beautifulsoup4==4.12.3
discord.py==2.6.4
python-dotenv==1.2.1
